Force $date variable to string. Fixes #153
The `mdjm_format_short_date()` function checks if the `$date` parameter
has been passed to it. If it has not it uses
`current_time('timestamp')` to define `$date`.

`current_time('timestamp')` returns the value as a double and therefore
our checks to determine if the variable is a unix timestamp are failing.

To correct this validation we need to use...

    if ( empty( $date ) )	{
	    $date = (string) current_time( 'timestamp' );
    }